Pushing ADC progress to myself

+++ embedded/include/current_sense.h
@@ -6,5 +6,5 @@
 #define CURRENT_SENSE_PIN 11
 
 int current_sense_init(void);
-uint32_t current_sense_read(void);
+double current_sense_read(void);
