+++ embedded/include/voltage_sense.h
@@ -6,5 +6,5 @@
 #define VOLTAGE_SENSE_PIN 10
 
 int voltage_sense_init(void);
-uint32_t voltage_sense_read(void);
+double voltage_sense_read(void);
